pyproject.toml
[tool.poetry]
name = "electo"
version = "0.1.0"
description = ""
authors = ["jakhongir <ganiyevuzb@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.3"
django-ckeditor = "^6.7.0"
pillow = "^10.0.0"
django-allauth = "^0.54.0"
django-environ = "^0.10.0"
django-redis = "^5.3.0"
celery = "^5.3.1"
model-bakery = "^1.13.0"
faker = "^19.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
